+++ src/components/select/components/focus-cover.js
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 
-export default function FocusCover({children}) {  
+export default function FocusCover({ children }) {  
   return ReactDOM.createPortal(
     children,
     document.body